# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FDA Regulatory Intelligence Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/summaries": {
      "post": {
        "tags": [
          "summaries"
        ],
        "summary": "Create Summary",
        "description": "Generate and persist a citation-bound grounded summary or refusal.",
        "operationId": "create_summary_api_summaries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResult"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/summaries/{summary_id}": {
      "get": {
        "tags": [
          "summaries"
        ],
        "summary": "Get Summary",
        "description": "Return a durable grounded summary owned by the caller.",
        "operationId": "get_summary_api_summaries__summary_id__get",
        "parameters": [
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResult"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "Draft",
          "Final",
          "Withdrawn"
        ],
        "title": "DocumentStatus",
        "description": "Lifecycle status of an FDA guidance document."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable, secret-free error message."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Correlation id for support/debugging."
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional field-level validation details."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Sanitized error detail returned to API clients."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope for all structured API error responses."
      },
      "EvidenceCard": {
        "properties": {
          "citation_id": {
            "type": "string",
            "title": "Citation Id",
            "description": "Stable answer-local citation marker, such as '[1]'."
          },
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Stable internal FDA document identifier."
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id",
            "description": "Stable source chunk identifier for exact passage lookup."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "FDA guidance title."
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id",
            "description": "Normalized section identifier when available."
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title",
            "description": "Human-readable section heading."
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number",
            "description": "PDF page number for source inspection."
          },
          "passage": {
            "type": "string",
            "title": "Passage",
            "description": "Exact quoted or cited source passage."
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url",
            "description": "FDA source URL."
          },
          "version_hash": {
            "type": "string",
            "title": "Version Hash",
            "description": "SHA-256 hash of the source document version."
          },
          "document_status": {
            "$ref": "#/components/schemas/DocumentStatus",
            "description": "Draft, Final, or Withdrawn."
          },
          "retrieval_score": {
            "type": "number",
            "title": "Retrieval Score",
            "description": "Dense/BM25/RRF retrieval score used for ranking."
          },
          "rerank_score": {
            "type": "number",
            "title": "Rerank Score",
            "description": "Cross-encoder reranker score."
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Final evidence confidence used by guardrails."
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At",
            "description": "Timestamp of retrieval for auditability."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "citation_id",
          "document_id",
          "chunk_id",
          "title",
          "passage",
          "source_url",
          "version_hash",
          "document_status",
          "retrieval_score",
          "rerank_score",
          "confidence"
        ],
        "title": "EvidenceCard",
        "description": "A verifiable citation binding a claim to an FDA source passage."
      },
...
      "SummaryRequest": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Stable internal FDA document identifier."
          },
          "summary_type": {
            "$ref": "#/components/schemas/SummaryType",
            "default": "summary"
          },
          "compare_version_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compare Version Hash",
            "description": "Prior version hash to diff against for key_changes summaries."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "SummaryRequest",
        "description": "A request to summarize a guidance document from grounded evidence."
      },
      "SummaryResult": {
        "properties": {
          "summary_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Id",
            "description": "Durable server-owned summary id."
          },
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Stable internal FDA document identifier."
          },
          "summary_type": {
            "$ref": "#/components/schemas/SummaryType"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Generated summary text with inline citation markers."
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceCard"
            },
            "type": "array",
            "title": "Evidence"
          },
          "refused": {
            "type": "boolean",
            "title": "Refused",
            "default": false
          },
          "refusal_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refusal Reason"
          },
          "current_version_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Hash"
          },
          "previous_version_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Version Hash"
          },
          "faithfulness_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faithfulness Passed"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "summary_type",
          "text"
        ],
        "title": "SummaryResult",
        "description": "A grounded summary/key-requirement/key-change result."
      },
      "SummaryType": {
        "type": "string",
        "enum": [
          "summary",
          "key_requirements",
          "key_changes"
        ],
        "title": "SummaryType",
        "description": "Kind of grounded summarization output."
      },
...
    }
  }
}